Add config and resources changes, move CurrentUserBean class
