+++ transmart-oauth/src/main/groovy/org/transmart/oauth/AuthorizationDecorator.groovy
@@ -1,4 +1,4 @@
-package org.transmart.authorization
+package org.transmart.oauth
 
 interface AuthorizationDecorator<T> {
