+++ transmartApp/src/main/groovy/org/transmart/authorization/QueriesResourceAuthorizationDecorator.groovy
@@ -5,6 +5,7 @@ import grails.util.Holders
 import groovy.util.logging.Slf4j
 import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.web.context.request.RequestContextHolder
+import org.transmart.oauth.AuthorizationDecorator
 import org.transmartproject.core.exceptions.AccessDeniedException
 import org.transmartproject.core.exceptions.InvalidRequestException
 import org.transmartproject.core.exceptions.NoSuchResourceException
